BUG/MINOR: halog: exit with error when -srv and -ic are set simultaneosly
Exit with error, when '-srv' and '-ic' output filters are set simultaneosly.

Otherwise, halog crashes. While processing input lines filter_count_ip() is
applied at first, if '-ic' is set. filter_count_srv_status() is never applied.
The logic seems to be exclusive for these two filters here:

	if (line_filter) {
		if (filter & FILT_COUNT_IP_COUNT)
			filter_count_ip(source_field, accept_field, time_field, &t);
		else if (filter2 & FILT2_EXTRACT_CAPTURE)
			filter_extract_capture(accept_field, time_field, filt2_capture_block, filt2_capture_field);
		else
			line_filter(accept_field, time_field, &t);

Due to this exclusivity, the ebtree with log records contains only the entries
corresponded to struct url_stat (allocated in filter_count_ip()). When output
is printed filter_count_srv_status() is applied at first. It takes the first
node from the ebtree and cast it via container_of to struct srv_st. So, any
further access to this srv_st can provoke heap overflow. This is reported in
the issue #3031 and can be reproduced with ASAN.

Code of halog is heavily optimized for performance. So, it's better to check
filters compatibility, while parsing the command line, instead of changing the
logic and the order in which the filters are applied.

This should be backported in all stable versions.